refac: seperate SideBar
