+++ src/components/common/header/HeaderMobile/styled.ts
@@ -1,5 +1,4 @@
-import { ButtonBase } from 'components/common/styled/elements';
-import { Column, Row } from 'components/common/styled/layout';
+import { Row } from 'components/common/styled/layout';
 import styled from 'styled-components';
 
 export const HeaderMobileWrapper = styled(Row)`
@@ -8,84 +7,18 @@ export const HeaderMobileWrapper = styled(Row)`
   }
 
   & > ${Row}.header-wrap {
-    padding: 18px;
-
-    width: 100%;
-    height: 32px;
-
-    align-items: center;
-    justify-content: space-between;
-  }
-
-  & > ${Column}.side-bar {
-    align-items: flex-end;
-
-    position: absolute;
-
-    top: 0;
-    right: 0;
-
-    background: white;
-
-    gap: 40px;
-    transition: 0.2s;
-
-    z-index: 1;
-
-    & > * {
-      margin-right: 20px;
-
-      font-family: 'Poppins-Bold';
-      font-size: 24px;
-      font-weight: bold;
-
-      white-space: pre;
-    }
-
-    &.open {
-      visibility: visible;
-      width: 195px;
-      height: 100vh;
-      opacity: 1;
-      padding-top: 63px;
-    }
-    &.close {
-      visibility: hidden;
-      width: 245px;
-      height: calc(100vh + 100px);
-      opacity: 0;
-    }
+    padding: 20px;
   }
 
   @media screen and (min-width: 768px) and (max-width: 1023px) {
     & > ${Row}.header-wrap {
-      padding: 30px;
+      padding: 45px 30px 0;
 
-      height: 32px;
-
-      & > div.logo {
-        color: white;
-      }
-    }
-
-    & > ${Column}.side-bar {
-      gap: 63px;
-
-      & > * {
-        margin-right: 30px;
-
-        font-size: 30px;
-      }
-
-      & > ${ButtonBase}.close-button {
-        margin-top: 80px;
-      }
-
-      &.open {
-        width: 417px;
-      }
-      &.close {
-        width: 245px;
+      & > a {
+        img {
+          width: 60px;
+          height: 30px;
+        }
       }
     }
   }
